# demo2kratos/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: demo2kratos-title
    version: 0.0.1
paths:
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /myhomepage/info:
        get:
            tags:
                - Myhomepage
            operationId: Myhomepage_GetMyInfo
            parameters:
                - name: account_name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.myhomepage.GetMyInfoReply'
        post:
            tags:
                - Myhomepage
            operationId: Myhomepage_SetMyInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.myhomepage.SetMyInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.myhomepage.SetMyInfoReply'
    /myhomepage/name/{account_name}:
        get:
            tags:
                - Myhomepage
            operationId: Myhomepage_GetMyName
            parameters:
                - name: account_name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.myhomepage.GetMyNameReply'
components:
    schemas:
        api.myhomepage.GetMyInfoReply:
            type: object
            properties:
                name:
                    type: string
                myFavoriteDevLanguage:
                    type: string
                english_name:
                    type: string
                chinese_name:
                    type: string
        api.myhomepage.GetMyNameReply:
            type: object
            properties:
                name:
                    type: string
        api.myhomepage.SetMyInfoReply:
            type: object
            properties:
                message:
                    type: string
        api.myhomepage.SetMyInfoRequest:
            type: object
            properties:
                name:
                    type: string
                myFavoriteDevLanguage:
                    type: string
                english_name:
                    type: string
                chinese_name:
                    type: string
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Greeter
      description: The greeting service definition.
    - name: Myhomepage